requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.2
mistralai>=0.0.12
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-chroma>=0.1.0
chromadb>=0.4.18
openai>=1.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
